Add files via upload
This project was the beginning and my introduction to Access Database. This is a phenomenal tool for storing, manipulating your information. In this project because I was working on a touch screen computer I had to activate the Touch Mode access on the Quick access tool bar, examine the fields and records respectively, identify the unique identifiers, rename fields, set the Primary Key and set each Data Type to the right column. Basically, I had to had new fields and saving various tables.